pyproject.toml
[project]
name = "python-threadsafe-serial"
description = "A serial port manager, allowing concurrent connections to a serial port instance"
version = "0.0.7"
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
authors = [
  { email = "matthew@davidson.engineering", name = "Matthew Davidson" },
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Operating System :: Microsoft :: Windows",
  "Operating System :: POSIX",
  "Operating System :: MacOS",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]
dependencies = ["pyserial>=3"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/threadsafe_serial"]

[dependency-groups]
dev = ["pytest>=7.1.1", "pytest-cov>=5.0.0", "coverage>=7.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["src/threadsafe_serial"]

[tool.coverage.report]
show_missing = true
skip_covered = true
